fix more builder
fix fix more builder I broke after sorting modules by dependency
